+++ examples/processTransaction.ts
@@ -147,7 +147,7 @@ async function main() {
             }
         }
     ];
-    const totalPrice = salesTicketResult[0].salePrice + salesTicketResult[0].salePrice;
+    const amount = salesTicketResult[0].salePrice + salesTicketResult[0].salePrice;
     const coaAuthorization = await sskts.service.transaction.placeOrder.createSeatReservationAuthorization(
         transactionId,
         individualScreeningEvent,
@@ -177,8 +177,8 @@ async function main() {
     const gmoAuthorization = await sskts.service.transaction.placeOrder.createCreditCardAuthorization(
         transactionId,
         orderId,
-        totalPrice,
-        '1',
+        amount,
+        sskts.GMO.utils.util.Method.Lump,
         {
             cardNo: '4111111111111111',
             expire: '2012',
@@ -198,8 +198,8 @@ async function main() {
     await sskts.service.transaction.placeOrder.createCreditCardAuthorization(
         transactionId,
         orderId,
-        totalPrice,
-        '1',
+        amount,
+        sskts.GMO.utils.util.Method.Lump,
         {
             cardNo: '4111111111111111',
             expire: '2012',
@@ -224,36 +224,6 @@ async function main() {
     debug('confirming transaction...');
     const order = await sskts.service.transaction.placeOrder.confirm(transactionId)(transactionAdapter);
     debug('confirmed', order);
-
-    // 照会してみる
-    // const key = sskts.factory.orderInquiryKey.create({
-    //     theaterCode: theaterCode,
-    //     orderNumber: coaAuthorization.result.tmp_reserve_num,
-    //     telephone: telephone
-    // });
-    // const inquiryResult = await sskts.service.order.makeInquiry(key)(orderAdapter);
-    // debug('makeInquiry result:', inquiryResult.get());
-
-    // メール追加
-    //     const content = `
-    // テスト 購入 ${profile.familyName} ${profile.givenName}様\n
-    // -------------------------------------------------------------------\n
-    // ◆購入番号 ：${coaAuthorization.result.tmp_reserve_num}\n
-    // ◆電話番号 ：${profile.telephone}\n
-    // ◆合計金額 ：${totalPrice}円\n
-    // -------------------------------------------------------------------\n
-    // シネマサンシャイン\n
-    // -------------------------------------------------------------------\n
-    // `;
-    // debug('adding email...');
-    // const notification = sskts.factory.notification.email.create({
-    //     from: 'noreply@example.net',
-    //     to: process.env.SSKTS_DEVELOPER_EMAIL,
-    //     subject: '購入完了',
-    //     content: content
-    // });
-    // await sskts.service.transaction.placeOrder.addEmail(transactionId, notification)(transactionAdapter);
-    // debug('email added.');
 
     redisClient.quit();
     mongoose.disconnect();